short video created for 9/2/2023

+++ Scrapers/videoMaker.py
@@ -6,6 +6,24 @@ from moviepy.video.io.ffmpeg_tools import ffmpeg_extract_subclip
 from moviepy.editor import VideoClip, TextClip, VideoFileClip, AudioFileClip, clips_array, concatenate_videoclips
 from fileDetails import get_mp3_length
 
+def splitTextForWrap(input_str: str, line_length: int):
+    words = input_str.split(" ")
+    line_count = 0
+    split_input = ""
+    for word in words:
+        line_count += 1
+        line_count += len(word)
+        if line_count > line_length:
+            split_input += "\n"
+            line_count = len(word) + 1
+            split_input += word
+            split_input += " "
+        else:
+            split_input += word
+            split_input += " "
+    
+    return split_input
+
 def randomVideoSegment(input_video_filepath, input_audio_filepath, output_video_filepath, duration):
     total_duration_seconds = 12 * 30 + 35
     # Generate a random start time within the valid range
@@ -26,7 +44,7 @@ def randomVideoSegment(input_video_filepath, input_audio_filepath, output_video_
     random_segment.write_videofile(output_video_filepath, codec="libx264")
 
 
-def textOverlay(video_path, text_input, characters_per_second, output_video_path):
+def textOverlay(video_path, text_input, output_video_path):
     # Get the current working directory of the script
     script_dir = os.path.dirname(os.path.abspath(__file__))
     # Construct the relative path to ffmpeg.exe
@@ -34,51 +52,74 @@ def textOverlay(video_path, text_input, characters_per_second, output_video_path
 
     start_time = 0
     video_segments = []  # To store paths of individual video segments
+    durations = []  # Initialize an empty list to store the durations
+    # Open the line_times file to fill line durations array
+    with open(f"{video_path.split('.')[0]}_line_times.txt", "r") as file:
+        for line in file:
+            parts = line.split()
+            duration_str = parts[-1]
+            try:
+                duration = float(duration_str)
+                durations.append(duration)
+            except ValueError:
+                print(f"Skipped invalid line: {line.strip()}")
+
+    duration_i = 0
     # Loop through the text_array and overlay text every 5 seconds
     for i, text in enumerate(text_input):
         print(text)
         # Remove leading and trailing whitespace and normal spaces from the line
         text = text.strip()
-        
         noWhitespaceText = text.replace(" ", "")
-        # Sleep for time it takes to read the sentence, adjust as needed
-        sleep_time = (len(noWhitespaceText) / characters_per_second)
-        if sleep_time == 0:
+        if not noWhitespaceText:
             continue
 
+        wrappedText = splitTextForWrap(text, 30)
         cmd = [
             ffmpeg_exe_path,
             "-nostdin",  # Disable interaction with standard input
             "-i", video_path,
-            "-vf", f"drawtext=text='{text}':x=10:y=10:fontsize=24:fontcolor=white:fontfile=C\\:/Windows/fonts/arial.ttf",
+            "-vf", f"drawtext=text='{wrappedText}':x=(w-text_w)/2:y=(h-text_h)/3:fontsize=55:fontcolor=white:fontfile=C\\:/Windows/fonts/arial.ttf:bordercolor=black:borderw=5:",
             "-c:a", "copy",
-            "-ss", str(start_time),
-            "-t", str(sleep_time),
+            "-ss", str(start_time),     
+            "-t", str(durations[duration_i]),
             "-y",
-            f"segment_{i}.mp4"  # Output path for this segment
+            f"segment_{duration_i}.mp4"  # Output path for this segment
         ]
-        print(start_time, " ", (start_time + sleep_time), " ", sleep_time)
+
+        print(start_time, " ", (start_time + durations[duration_i]), " ", durations[duration_i])
         subprocess.run(cmd, stdout=subprocess.PIPE, stderr=subprocess.PIPE, text=True)
         
         # Append the path of the generated segment to the list
-        video_segments.append(f"segment_{i}.mp4")
+        video_segments.append(f"segment_{duration_i}.mp4")
 
         # Update start time for the next segment
-        start_time += sleep_time
+        start_time += durations[duration_i]
+        duration_i += 1
 
     # Load the video clips
-    video_clips = [VideoFileClip(path) for path in video_segments]
-    # Concatenate the video clips sequentially
+    # List to store video clips
+    video_clips = []
+    # Iterate through the list of segment paths
+    for path in video_segments:
+        if os.path.exists(path):
+            clip = VideoFileClip(path)
+            video_clips.append(clip)
+        else:
+            # Handle non-existent files (you can print a message or take other actions)
+            print(f"File not found: {path}")
+    # Concatenate the video clips sequentially; maybe change compose to chain
     final_video = concatenate_videoclips(video_clips, method="compose")
     # Write the concatenated video to a file
-    final_video.write_videofile(output_video_path, codec="libx264")
+    final_video.write_videofile(output_video_path, codec="libx264", threads=8)
     # Close the video clips
     for clip in video_clips:
         clip.close()
 
     # Clean up individual video segments
     for segment in video_segments:
-        os.remove(segment)
+        if os.path.exists(segment):
+            os.remove(segment)
 
     print("Overlay complete.")
 
@@ -87,59 +128,35 @@ if __name__ == "__main__":
     background_video_path = "SubwaySurfers/subwaySurfers.mp4"
 
     folder_path = "RedditPosts/2023-09-02/Texts"
-    for subreddit in os.listdir(folder_path):
-        post_path = f"{folder_path}/{subreddit}"
-        for post in os.listdir(post_path):
-            if post.endswith(".mp3"):
-                mp3_file_path = f"{post_path}/{post}"
-                output_video_path = f"{post_path}/{post.split('.')[0]}.mp4"
-                duration = get_mp3_length(mp3_file_path)
-                randomVideoSegment(background_video_path, mp3_file_path, output_video_path, duration)
+    # for subreddit in os.listdir(folder_path):
+    #     post_path = f"{folder_path}/{subreddit}"
+    #     for post in os.listdir(post_path):
+    #         if post.endswith(".mp3"):
+    #             mp3_file_path = f"{post_path}/{post}"
+    #             output_video_path = f"{post_path}/{post.split('.')[0]}.mp4"
+    #             duration = get_mp3_length(mp3_file_path)
+    #             randomVideoSegment(background_video_path, mp3_file_path, output_video_path, duration)
     
 
     for subreddit in os.listdir(folder_path):
         post_path = f"{folder_path}/{subreddit}"
         for post in os.listdir(post_path):
-            if post.endswith(".mp4"):
+            if post.endswith(".mp4") and not post.endswith("F.mp4"):
                 mp4_file_path = f"{post_path}/{post}"
                 mp4_output_path = f"{post_path}/{post.split('.')[0]}F.mp4"
                 text_input = []
                 non_empty_char_count = 0
-                text_file_path = f"{post_path}/{post.split('.')}.txt"
+                text_file_path = f"{post_path}/{post.split('.')[0]}.txt"
                 # Open the file for reading
                 with open(text_file_path, 'r', encoding='utf-8') as file:
                     # Read each line and add it to the list
                     for line in file:
                         # Remove leading and trailing whitespace from the line
                         line = line.strip()
-                        text_input.append(line)
+                        removed_spaces_line = line.replace(" ", "")  # This removes spaces
+                        if removed_spaces_line:
+                            non_empty_char_count += len(removed_spaces_line)
+                            text_input.append(line)
+                            
+                textOverlay(mp4_file_path, text_input, mp4_output_path)
 
-                        # Remove spaces from the line and count non-empty characters
-                        line = line.replace(" ", "")  # This removes spaces
-                        if line:
-                            non_empty_char_count += len(line)
-                    characters_per_second = non_empty_char_count / get_mp3_length(mp3_file_path)
-                textOverlay(mp4_file_path, text_input, characters_per_second, mp4_output_path)
-
-    # mp4_file_path = "AmItheAsshole1.mp4"
-    # mp4_output_path = "AmItheAsshole1F.mp4"
-    # mp3_file_path = "AmItheAsshole1.mp3"
-    # text_file_path = "AmItheAsshole1.txt"
-    
-    # non_empty_char_count = 0
-    # # Open the file for reading
-    # with open(text_file_path, 'r', encoding='utf-8') as file:
-    #     # Read each line and add it to the list
-    #     for line in file:
-    #         # Remove leading and trailing whitespace from the line
-    #         line = line.strip()
-    #         text_input.append(line)
-
-    #         # Remove spaces from the line and count non-empty characters
-    #         line = line.replace(" ", "")  # This removes spaces
-    #         if line:
-    #             non_empty_char_count += len(line)
-
-    # characters_per_second = non_empty_char_count / get_mp3_length(mp3_file_path)
-
-    # textOverlay(mp4_file_path, text_input, characters_per_second, mp4_output_path)
